Put hooks in individual folders + set src as root
