+++ ui/tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "baseUrl": "src",
     "outDir": "./dist/",
     "sourceMap": true,
     "strictNullChecks": true,
@@ -23,6 +24,6 @@
     "allowJs": true
   },
   "include": [
-    "./src/"
+    "src"
   ]
 }
